+++ run/kubejs/server_scripts/recipes/firmalife.js
@@ -1,52 +1,54 @@
 // requires firmalife
 
-function firmalife(f) {
-    f.drying(
+ServerEvents.recipes(e => {
+    let { firmalife } = e.recipes;
+
+    firmalife.drying(
         'minecraft:dirt',
         'minecraft:stone'
     ).id('kubejs:drying');
 
-    f.smoking(
+    firmalife.smoking(
         'minecraft:stone',
         'minecraft:deepslate'
     ).id('kubejs:smoking');
 
-    f.mixing_bowl()
+    firmalife.mixing_bowl()
         .outputItem('5x minecraft:grass')
         .fluidIngredient(Fluid.of('minecraft:water', 1000))
         .id('kubejs:mixing_bowl');
 
-    f.oven('minecraft:oak_log', 3, 3)
+    firmalife.oven('minecraft:oak_log', 3, 3)
         .id('kubejs:oven_0');
-    f.oven('minecraft:spruce_log', 3, 3, 'minecraft:oak_log')
+    firmalife.oven('minecraft:spruce_log', 3, 3, 'minecraft:oak_log')
         .id('kubejs:oven_1');
 
-    f.stinky_soup(['#minecraft:flowers'], Fluid.of('minecraft:water', 50), 5, 5)
+    firmalife.stinky_soup(['#minecraft:flowers'], Fluid.of('minecraft:water', 50), 5, 5)
         .id('kubejs:stinky_soup');
 
-    f.vat()
+    firmalife.vat()
         .outputItem('minecraft:dirt')
         .inputFluid(Fluid.of('minecraft:lava', 500))
         .length(60)
         .id('kubejs:vat');
 
-    f.stomping(
+    firmalife.stomping(
         'minecraft:dirt',
         'minecraft:stone',
         'tfc:block/charcoal_pile',
         'tfc:block/powder/charcoal',
-        'tfc:block/charcoal/fall'
+        'tfc:block.charcoal.fall'
     ).id('kubejs:stomping');
 
-    f.press(
+    firmalife.press(
         'minecraft:smooth_stone',
         'minecraft:cobblestone',
         'tfc:block/charcoal_pile',
         'tfc:block/powder/charcoal',
-        'tfc:block/charcoal/fall'
+        'tfc:block.charcoal.fall'
     ).id('kubejs:press');
 
-    f.bowl_pot('minecraft:cooked_beef', [
+    firmalife.bowl_pot('minecraft:cooked_beef', [
         'minecraft:dirt',
         'minecraft:stone'
     ], 'minecraft:water', 20, 100, food => {
@@ -54,7 +56,7 @@ function firmalife(f) {
         food.saturation(2);
         food.protein(3);
     }).id('kubejs:bowl_pot_0');
-    f.bowl_pot('tfc:food/red_apple', [
+    firmalife.bowl_pot('tfc:food/red_apple', [
         'minecraft:poppy',
         'minecraft:oak_log'
     ], 'minecraft:lava', 20, 100, {
@@ -62,10 +64,4 @@ function firmalife(f) {
         fruit: 2,
         decay_modifier: 0.9
     }).id('kubejs:bowl_pot_1');
-}
-
-ServerEvents.recipes(e => {
-    let { firmalife } = e.recipes;
-
-    firmalife(firmalife);
 })